Add fetchGroup and fetchGroups ducks
