+++ static/js/containers/GroupContainer.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 import * as Action from '../actions';
+import * as fetchGroupActions from '../ducks/fetchGroup';
 import Group from '../components/Group';
 
 class GroupContainer extends React.Component {
@@ -51,7 +52,7 @@ const mapDispatchToProps = (dispatch, ownProps) => (
       Action.deleteGroupUser({ username, group_id })
     ),
     fetchGroup: id => dispatch(
-      Action.fetchGroup(id)
+      fetchGroupActions.fetchGroup(id)
     )
   }
 );
